# backend/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Invitation:
    properties:
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      expiresAt:
        type: string
      id:
        type: integer
      restaurant:
        $ref: '#/definitions/models.Restaurant'
      restaurantID:
        type: integer
      role:
        type: string
      sender:
        $ref: '#/definitions/models.User'
      senderID:
        type: integer
      status:
        $ref: '#/definitions/models.InvitationStatus'
      token:
        type: string
      updatedAt:
        type: string
    type: object
  models.InvitationStatus:
    enum:
    - PENDING
    - ACCEPTED
    - REJECTED
    - EXPIRED
    type: string
    x-enum-varnames:
    - InvitationPending
    - InvitationAccepted
    - InvitationRejected
    - InvitationExpired
  models.Merchant:
    properties:
      business_name:
        description: Nom de l'entreprise
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email_pro:
        description: Email valide requis
        type: string
      id:
        type: integer
      phone_number:
        description: Numéro de téléphone (optionnel, max 15 caractères)
        type: string
      siret:
        description: Numéro SIRET
        type: string
      updatedAt:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relation vers User (clé étrangère avec cascade)
      user_id:
        description: Relation 1 à 1 vers User
        type: integer
    required:
    - business_name
    - email_pro
    - siret
    type: object
...
  models.Restaurant:
    properties:
      address:
        description: Adresse complète
        type: string
      category:
        allOf:
        - $ref: '#/definitions/models.Category'
        description: Relation avec Category (clé étrangère)
      category_id:
        description: ID de la catégorie (clé étrangère)
        type: integer
      city:
        description: Ville
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      merchant:
        allOf:
        - $ref: '#/definitions/models.Merchant'
        description: Relation avec le commerçant
      merchant_id:
        description: ID du commerçant (clé étrangère)
        type: integer
      name:
        description: Nom du restaurant (obligatoire)
        type: string
      phone_number:
        description: Numéro de téléphone (optionnel, max 15 caractères)
        type: string
      postal_code:
        description: Code postal (limité à 10 caractères pour compatibilité internationale)
        type: string
      rating:
        description: Note moyenne (sur 5)
        type: number
      siren:
        description: SIREN (exactement 9 chiffres, unique)
        type: string
      updatedAt:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: Validation d'email
        type: string
      full_name:
        description: Nom complet requis
        type: string
      id:
        type: integer
      is_admin:
        description: Est-ce un administrateur ?
        type: boolean
      password_hash:
        description: Hash du mot de passe
        type: string
      phone:
        description: Téléphone requis
        type: string
      updatedAt:
        type: string
    required:
    - email
    - full_name
    - password_hash
    - phone
    type: object
...
info:
  contact: {}
paths:
...
  /api/invitations:
    post:
      consumes:
      - application/json
      description: Permet à un marchand d'inviter une personne à rejoindre son restaurant
        en tant que staff
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Informations de l'invitation
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
            restaurant_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Invitation envoyée avec succès
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
        "400":
          description: Erreur dans la requête
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Non authentifié
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Accès non autorisé
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Envoyer une invitation à rejoindre un restaurant
      tags:
      - invitations
  /api/invitations/{invitationId}:
    delete:
      description: Permet à un marchand d'annuler une invitation envoyée
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de l'invitation
        in: path
        name: invitationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Invitation annulée avec succès
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Erreur dans la requête
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Non authentifié
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Annuler une invitation
      tags:
      - invitations
  /api/invitations/accept:
    get:
      consumes:
      - application/json
      description: Permet à un utilisateur d'accepter une invitation à rejoindre un
        restaurant en tant que staff
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Code d'invitation
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invitation acceptée avec succès
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Erreur dans la requête
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Non authentifié
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Accepter une invitation à rejoindre un restaurant
      tags:
      - invitations
...
  /api/restaurants/{restaurantId}/invitations:
    get:
      description: Permet à un marchand de voir toutes les invitations en attente
        pour son restaurant
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID du restaurant
        in: path
        name: restaurantId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Liste des invitations en attente
          schema:
            items:
              $ref: '#/definitions/models.Invitation'
            type: array
        "400":
          description: Erreur dans la requête
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Non authentifié
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Accès non autorisé
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Récupérer les invitations en attente pour un restaurant
      tags:
      - invitations
swagger: "2.0"